Bajs
git-svn-id: file:///home/srv/www/butt.se/pnp/svn/trunk@142 f672c3ff-8b41-4f22-a4ab-2adcb4f732c7

+++ CameraAPI/BarrelCorrection.h
@@ -21,8 +21,8 @@ class BarrelCorrection : public Filter
 		void calculatePixelMapping();
 		void solveLinearEquation(float M[8][9]);
 		
-		float m_distortedX[8];
-		float m_distortedY[8];
+		unsigned int m_distortedX[8];
+		unsigned int m_distortedY[8];
 		unsigned int m_outputWidth;
 		unsigned int m_outputHeight;
 		unsigned int *m_pixelMapping;
